requirements.txt
dfcx-scrapi
google-cloud-dialogflow-cx
google-auth
google-oauth
oauth2client
pyparsing==2.4.7
pandas
tabulate
gspread==5.10.0
gspread_dataframe
numpy
requests
pylint==2.8.3
pytest==6.0.2
pytest-cov==2.11.1
pytest-xdist==2.1.0
pyyaml==5.4
torch
transformers
sentencepiece
